Milestone1 achieved
Now the application can connect to a elm327 device and
send command to elm327 to get vehicle rpm/speed etc data.

I can also set the Gauge with rpm data. but there is a minor
issue that when set rpm in a thread, the gauge can't be dynamically
refreshed. need to debug.

Signed-off-by: Oak Zeng <zengshanjun@gmail.com>